+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ v1.1.2
 *In development*
 
 - Baseline name property can contain *{Version}* to add actual version capability value to the baseline name
+- New config property 'gecko_driver_path' in [Browser] section to configure the Gecko/Marionette driver location
 
 v1.1.1
 ------